Rewrite gateway to run in asyncio loop
