+++ docker-compose.yml
@@ -35,6 +35,8 @@ services:
     environment:
       SECRET_KEY: 98dn7ncq39up4c_2fRVQ
       USERS_API_BASE_URL: "http://users:8080"
+      GATEWAY_HOST: "0.0.0.0"
+      GATEWAY_PORT: 8080
     networks:
       - gateway_net
       - frontend
